Updated YAML v2, corrected messages & tickets
This commit updates the YAML v2 to include some of the new changes in the final API implementation. However, the YAML will require further changes as it is still incomplete. Also, this commit makes some slight corrections to the tickets API controller (specifically, it handles some input errors) and the comments API controller (specifically, it corrects an erroneous HTTP return code put in earlier).